Split source in include and src. Fixed error in merge of libs.
