+++ CMakeLists.txt
@@ -51,73 +51,16 @@ endif()
 
 set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Wall -Wextra")
 
-# Enable support for external projects
-include(ExternalProject)
-
-########
-# mosquitto
-ExternalProject_Add(mosquitto
-	SOURCE_DIR ${PROJECT_SOURCE_DIR}/vendor/mosquitto-1.3.5
-	CMAKE_ARGS -DWITH_SRV=OFF -DWITH_TLS=OFF -DCMAKE_INSTALL_PREFIX=<INSTALL_DIR> -DCMAKE_C_FLAGS=-fpic -DCMAKE_CXX_FLAGS=-fpic)
-
-# Store mosquitto variables
-ExternalProject_Get_Property(mosquitto install_dir)
-set(MOSQUITTO_LIBRARIES ${install_dir}/lib/libmosquittopp.a ${install_dir}/lib/libmosquitto.a) 
-set(MOSQUITTO_INCLUDE_DIR "${install_dir}/include")
-
-# Install headers
-install(DIRECTORY "${MOSQUITTO_INCLUDE_DIR}/" DESTINATION include/fast-lib/external)
-########
-
-########
-# Boost
-include(add_boost)
-include_directories(SYSTEM ${BoostSourceDir})
-
-# Install headers
-install(DIRECTORY "${BoostSourceDir}/boost" DESTINATION include/fast-lib/external)
-########
-
-########
-# yaml-cpp
-ExternalProject_Add(yaml-cpp
-	SOURCE_DIR ${PROJECT_SOURCE_DIR}/vendor/yaml-cpp-0.5.1
-	CMAKE_ARGS -DCMAKE_CXX_FLAGS=-fpic -DBOOST_ROOT=${BoostSourceDir} -DBUILD_SHARED_LIBS=OFF -DCMAKE_INSTALL_PREFIX=<INSTALL_DIR>)
-
-# Store yaml-cpp variables
-ExternalProject_Get_Property(yaml-cpp install_dir)
-set(YAMLCPP_LIBRARY "${install_dir}/lib/libyaml-cpp.a")
-set(YAMLCPP_INCLUDE_DIR "${install_dir}/include")
-
-# Install headers
-install(DIRECTORY "${YAMLCPP_INCLUDE_DIR}/" DESTINATION include/fast-lib/external)
-########
-
-# Add support for merging static libraries
-include(MergeStaticLibraries)
-
 ### Define header, source and lib files
 # Headers
-set(HEADERS_MESSAGE_AGENT
-	"${CMAKE_CURRENT_SOURCE_DIR}/src/message/agent/init.hpp"
-	"${CMAKE_CURRENT_SOURCE_DIR}/src/message/agent/init_agent.hpp"
-	"${CMAKE_CURRENT_SOURCE_DIR}/src/message/agent/stop_monitor.hpp"
-)
-
-#set(HEADERS_MESSAGE_MIGFRA
-#)
-
-set(HEADERS_ROOT
-	"${CMAKE_CURRENT_SOURCE_DIR}/src/communicator.hpp"
-	"${CMAKE_CURRENT_SOURCE_DIR}/src/mqtt_communicator.hpp"
-	"${CMAKE_CURRENT_SOURCE_DIR}/src/serializable.hpp"
-)
-
-
+include_directories("${CMAKE_CURRENT_SOURCE_DIR}/include")
 set(HEADERS
-	"${HEADERS_ROOT}"
-	"${HEADERS_MESSAGE_AGENT}"
-#	"${HEADERS_MESSAGE_MIGFRA}"
+	"${CMAKE_CURRENT_SOURCE_DIR}/include/fast-lib/communicator.hpp"
+	"${CMAKE_CURRENT_SOURCE_DIR}/include/fast-lib/mqtt_communicator.hpp"
+	"${CMAKE_CURRENT_SOURCE_DIR}/include/fast-lib/serializable.hpp"
+	"${CMAKE_CURRENT_SOURCE_DIR}/include/fast-lib/message/agent/init.hpp"
+	"${CMAKE_CURRENT_SOURCE_DIR}/include/fast-lib/message/agent/init_agent.hpp"
+	"${CMAKE_CURRENT_SOURCE_DIR}/include/fast-lib/message/agent/stop_monitor.hpp"
 )
 
 # Source
@@ -127,32 +70,82 @@ set(SRC
 	"${CMAKE_CURRENT_SOURCE_DIR}/src/message/agent/init.cpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/src/message/agent/init_agent.cpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/src/message/agent/stop_monitor.cpp"
-	"${HEADERS}"
 )
 
-# Include directories
-include_directories(SYSTEM "${MOSQUITTO_INCLUDE_DIR}")
-include_directories(SYSTEM "${YAMLCPP_INCLUDE_DIR}")
-
-### Build and installation targets
 # Add library
-add_library(${FASTLIB} ${SRC})
+add_library(${FASTLIB} "${SRC}" "${HEADERS}")
 
-# Merge with dependent libraries
-add_dependencies(${FASTLIB} ${BoostRegexLibs})
+# Enable support for external projects
+include(ExternalProject)
+
+########
+# Boost
+include(add_boost)
+
 get_property(BOOST_REGEX_LIBRARY TARGET ${BoostRegexLibs} PROPERTY LOCATION)
-set(LIBS ${MOSQUITTO_LIBRARIES} ${YAMLCPP_LIBRARY} ${BOOST_REGEX_LIBRARY})
+add_dependencies(${FASTLIB} "${BoostRegexLibs}")
+set(LIBS "${BOOST_REGEX_LIBRARY}")
 if(ENABLE_LOGGING)
-	add_dependencies(${FASTLIB} ${BoostLogLibs})
+	add_dependencies(${FASTLIB} "${BoostLogLibs}")
 	get_property(BOOST_LOG_LIBRARY TARGET ${BoostLogLibs} PROPERTY LOCATION)
 	get_property(BOOST_CHRONO_LIBRARY TARGET ${BoostChronoLibs} PROPERTY LOCATION) # dependency of boost.log
 	get_property(BOOST_DATE_TIME_LIBRARY TARGET ${BoostDateTimeLibs} PROPERTY LOCATION) # dependency of boost.log
 	get_property(BOOST_FILESYSTEM_LIBRARY TARGET ${BoostFilesystemLibs} PROPERTY LOCATION) # dependency of boost.log
 	get_property(BOOST_THREAD_LIBRARY TARGET ${BoostThreadLibs} PROPERTY LOCATION) # dependency of boost.log
 	get_property(BOOST_SYSTEM_LIBRARY TARGET ${BoostSystemLibs} PROPERTY LOCATION) # dependency of boost.log
-	set(LIBS ${LIBS} ${BOOST_LOG_LIBRARY} ${BOOST_CHRONO_LIBRARY} ${BOOST_DATE_TIME_LIBRARY} ${BOOST_FILESYSTEM_LIBRARY} ${BOOST_THREAD_LIBRARY} ${BOOST_SYSTEM_LIBRARY})
+	set(LIBS "${LIBS}" "${BOOST_LOG_LIBRARY}" "${BOOST_CHRONO_LIBRARY}" "${BOOST_DATE_TIME_LIBRARY}" "${BOOST_FILESYSTEM_LIBRARY}" "${BOOST_THREAD_LIBRARY}" "${BOOST_SYSTEM_LIBRARY}")
 endif()
+
+include_directories(SYSTEM ${BoostSourceDir})
+
+# Install headers
+install(DIRECTORY "${BoostSourceDir}/boost" DESTINATION include/fast-lib/external)
+########
+
+########
+# mosquitto
+ExternalProject_Add(mosquitto
+	SOURCE_DIR ${PROJECT_SOURCE_DIR}/vendor/mosquitto-1.3.5
+	CMAKE_ARGS -DWITH_SRV=OFF -DWITH_TLS=OFF -DCMAKE_INSTALL_PREFIX=<INSTALL_DIR> -DCMAKE_C_FLAGS=-fpic -DCMAKE_CXX_FLAGS=-fpic)
+
+add_dependencies(${FASTLIB} "mosquitto")
+
+# Store mosquitto variables
+ExternalProject_Get_Property(mosquitto install_dir)
+set(MOSQUITTO_LIBRARIES ${install_dir}/lib/libmosquittopp.a ${install_dir}/lib/libmosquitto.a) 
+set(MOSQUITTO_INCLUDE_DIR "${install_dir}/include")
+set(LIBS "${LIBS}" "${MOSQUITTO_LIBRARIES}")
+
+include_directories(SYSTEM "${MOSQUITTO_INCLUDE_DIR}")
+
+# Install headers
+install(DIRECTORY "${MOSQUITTO_INCLUDE_DIR}/" DESTINATION include/fast-lib/external)
+########
+
+########
+# yaml-cpp
+ExternalProject_Add(yaml-cpp
+	SOURCE_DIR ${PROJECT_SOURCE_DIR}/vendor/yaml-cpp-0.5.1
+	CMAKE_ARGS -DCMAKE_CXX_FLAGS=-fpic -DBOOST_ROOT=${BoostSourceDir} -DBUILD_SHARED_LIBS=OFF -DCMAKE_INSTALL_PREFIX=<INSTALL_DIR>)
+
+add_dependencies(${FASTLIB} "yaml-cpp")
+
+# Store yaml-cpp variables
+ExternalProject_Get_Property(yaml-cpp install_dir)
+set(YAMLCPP_LIBRARY "${install_dir}/lib/libyaml-cpp.a")
+set(YAMLCPP_INCLUDE_DIR "${install_dir}/include")
+set(LIBS "${LIBS}" "${YAMLCPP_LIBRARY}")
+
+include_directories(SYSTEM "${YAMLCPP_INCLUDE_DIR}")
+
+# Install headers
+install(DIRECTORY "${YAMLCPP_INCLUDE_DIR}/" DESTINATION include/fast-lib/external)
+########
+
+# Merge with dependent libraries
+include(MergeStaticLibraries)
 MERGE_STATIC_LIBRARIES(${FASTLIB} ALL "${LIBS}")
+
 # As librt should always be available it is linked dynamically and therefore has to be linked in applications
 # that use fast-lib as well.
 target_link_libraries(${FASTLIB} -lrt)
@@ -161,12 +154,8 @@ target_link_libraries(${FASTLIB} -lrt)
 install(TARGETS "${FASTLIB}"
 	LIBRARY DESTINATION "lib"
 	ARCHIVE DESTINATION "lib")
-install(FILES "${HEADERS_ROOT}"
-	DESTINATION "include/fast-lib")
-install(FILES "${HEADERS_MESSAGE_AGENT}"
-	DESTINATION "include/fast-lib/message/agent")
-#install(FILES "${HEADERS_MESSAGE_MIGFRA}"
-#	DESTINATION "include/fast-lib/message/migfra")
+install(DIRECTORY "include/"
+	DESTINATION "include")
 
 # Testing
 if(BUILD_EXAMPLES)
