+++ examples/comm_serial/CMakeLists.txt
@@ -12,7 +12,6 @@ set(FASTLIB_EXAMPLE_COMM_SERIAL "fastlib_example_comm_serial")
 set(SRC ${CMAKE_CURRENT_SOURCE_DIR}/main.cpp)
 
 # Include directories
-include_directories(${PROJECT_SOURCE_DIR}/src)
 include_directories(SYSTEM "${MOSQUITTO_INCLUDE_DIR}")
 include_directories(SYSTEM "${YAMLCPP_INCLUDE_DIR}")
